requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
alembic
pytest
google-generativeai
